Enhance stock dispatch management by updating access policies to allow elevated dispatch managers to edit and delete manual dispatch records. Refactor `ManualStockDispatchService` to incorporate new permissions and improve error handling. Update SQL migration to ensure correct data types for foreign keys. Revise frontend components to reflect changes in dispatch permissions and improve user experience with detailed loading states and error messages.

+++ backend/smart-erp/src/main/resources/db/migration/V35__stockdispatch_lines_soft_delete.sql
@@ -3,8 +3,8 @@
 
 CREATE TABLE IF NOT EXISTS stockdispatch_lines (
     id           BIGSERIAL PRIMARY KEY,
-    dispatch_id  BIGINT NOT NULL REFERENCES stockdispatches (id) ON DELETE CASCADE,
-    inventory_id BIGINT NOT NULL REFERENCES inventory (id),
+    dispatch_id  INTEGER NOT NULL REFERENCES stockdispatches (id) ON DELETE CASCADE,
+    inventory_id INTEGER NOT NULL REFERENCES inventory (id),
     quantity     INTEGER NOT NULL CHECK (quantity > 0),
     CONSTRAINT uq_stockdispatch_line UNIQUE (dispatch_id, inventory_id)
 );
